Moving to Apache Derby from sqlite.
* lib/sqlitejdbc-v044-nested.jar: Jar deleted.

* .../gotefarm/client/Characters.java,
  .../gotefarm/client/GoteFarmRPC.java,
  .../gotefarm/client/GoteFarmRPCAsync.java,
  .../gotefarm/client/JSCharRole.java,
  .../gotefarm/client/JSCharacter.java,
  .../gotefarm/client/JSEventSchedule.java,
  .../gotefarm/client/JSEventTemplate.java,
  .../giftoftheembalmer/gotefarm/client/JSRole.java,
  .../gotefarm/server/dao/GoteFarmDaoT.scala,
  .../gotefarm/server/dao/GoteFarmJdbcDao.scala,
  .../server/service/GoteFarmServiceImpl.scala,
  .../gotefarm/server/service/GoteFarmServiceT.scala,
  .../gotefarm/server/servlet/GoteFarmRPCImpl.scala:
  Use longs for primary keys.

* .../gotefarm/server/dao/GoteFarmJdbcDao.scala:
  Update SQL for Derby.

  (tableExists): Method deleted.

  (generateTables): Derby does not support "IF NOT EXISTS" like sqlite,
  instead handle schema changes with a simple migration system, keeping the
  state of the latest schema changes applied to the database in a new table
  schema_vers.
  Add proper foreign key constraints to all tables that need them, now that
  they are available.
  Name all constraints to get useful errors from constraint violations.
  Use CHAR(1) with 'Y' or 'N' values to store boolean values.

  (migrate_to_vers): If the passed in version is newer, run the DDL commands.
  Note: if one or more DDL commands fail, some may succeed and are not rolled
  back, leaving the database in a half-migrated state.  The database should be
  backed up before performing the migration, so recovery is possible.  The DDL
  commands should be tested for correctness before running on a production
  database.

* web/WEB-INF/applicationContext.xml
  (dataSource):  Update driverClass and jdbcUrl for Derby.

+++ src/com/giftoftheembalmer/gotefarm/client/Characters.java
@@ -162,9 +162,9 @@ public class Characters extends Composite {
         public void addCharacter() {
             errmsg.setText("");
 
-            GoteFarm.testService.newCharacter(GoteFarm.sessionID, realm.getText(), character.getText(), new AsyncCallback<Integer>() {
-                public void onSuccess(Integer result) {
-                    GoteFarm.testService.getCharacter(GoteFarm.sessionID, result.intValue(), new AsyncCallback<JSCharacter>() {
+            GoteFarm.testService.newCharacter(GoteFarm.sessionID, realm.getText(), character.getText(), new AsyncCallback<Long>() {
+                public void onSuccess(Long result) {
+                    GoteFarm.testService.getCharacter(GoteFarm.sessionID, result.longValue(), new AsyncCallback<JSCharacter>() {
                         public void onSuccess(JSCharacter chr) {
                             chrpanel.add(new Character(chr));
                         }
