+++ src/com/giftoftheembalmer/gotefarm/client/JSCharRole.java
@@ -3,9 +3,9 @@ package com.giftoftheembalmer.gotefarm.client;
 import com.google.gwt.user.client.rpc.IsSerializable;
 
 public class JSCharRole implements IsSerializable {
-    public int chrroleid;
-    public int cid;
-    public int roleid;
+    public long chrroleid;
+    public long cid;
+    public long roleid;
     public boolean approved;
 
     public JSCharRole() {
