+++ src/com/giftoftheembalmer/gotefarm/client/JSRole.java
@@ -3,7 +3,7 @@ package com.giftoftheembalmer.gotefarm.client;
 import com.google.gwt.user.client.rpc.IsSerializable;
 
 public class JSRole implements IsSerializable {
-    public int roleid;
+    public long roleid;
     public String name;
     public boolean restricted;
