+++ src/com/giftoftheembalmer/gotefarm/server/service/GoteFarmServiceT.scala
@@ -16,34 +16,34 @@ trait GoteFarmServiceT {
   def generateTables(): Unit
 
   @throws(classOf[InvalidCredentialsError])
-  def login(username: String, password: String): Int
+  def login(username: String, password: String): Long
 
   @throws(classOf[AlreadyExistsError])
-  def newUser(username: String, email: String, password: String): Int
+  def newUser(username: String, email: String, password: String): Long
 
   @throws(classOf[AlreadyExistsError])
   @throws(classOf[NotFoundError])
-  def newCharacter(uid: Int, realm: String, character: String): Int
+  def newCharacter(uid: Long, realm: String, character: String): Long
 
-  def getCharacters(uid: Int): List[JSCharacter]
+  def getCharacters(uid: Long): List[JSCharacter]
 
   @throws(classOf[NotFoundError])
-  def getCharacter(cid: Int): JSCharacter
+  def getCharacter(cid: Long): JSCharacter
 
   def getRoles: List[JSRole]
   @throws(classOf[AlreadyExistsError])
-  def addRole(name: String, restricted: Boolean): Int
+  def addRole(name: String, restricted: Boolean): Long
 
   def getBadges: List[String]
   @throws(classOf[AlreadyExistsError])
-  def addBadge(name: String, score: Int): Int
+  def addBadge(name: String, score: Int): Long
 
   @throws(classOf[AlreadyExistsError])
-  def addInstance(name: String): Int
+  def addInstance(name: String): Long
 
   @throws(classOf[AlreadyExistsError])
   @throws(classOf[NotFoundError])
-  def addBoss(instance: String, name: String): Int
+  def addBoss(instance: String, name: String): Long
 
   def getInstances: List[String]
 
@@ -54,7 +54,7 @@ trait GoteFarmServiceT {
   def getEventTemplates: List[String]
   @throws(classOf[AlreadyExistsError])
   @throws(classOf[NotFoundError])
-  def saveEventTemplate(et: JSEventTemplate): Int
+  def saveEventTemplate(et: JSEventTemplate): Long
 
   def getEventSchedules(name: String): List[JSEventSchedule]
 }
